requirements_dev.txt
autopep8==2.0.2
cssbeautifier==1.14.7
djlint==1.32.1
jsbeautifier==1.14.7
pycodestyle==2.10.0
pylint==2.17.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
selenium==4.9.0
